`diff_map` now uses `diffnd_map`. Move timing for `diff_map` to time/diffs.jl.
